feat(relay): improve token calculation and cost accuracy
Changes token types to int64 for larger values.
Refactors token calculation logic for better accuracy.
Updates cost calculation to use corrected token counts.
Adds fallback for calculating tokens when response is unavailable.

+++ internal/model/log.go
@@ -30,8 +30,8 @@ type RelayLog struct {
 	ChannelId        int               `json:"channel"`                                  // 实际使用的渠道ID
 	ChannelName      string            `json:"channel_name"`                             // 渠道名称
 	ActualModelName  string            `json:"actual_model_name"`                        // 实际使用模型名称
-	InputTokens      int               `json:"input_tokens"`                             // 输入Token
-	OutputTokens     int               `json:"output_tokens"`                            // 输出 Token
+	InputTokens      int64             `json:"input_tokens"`                             // 输入Token
+	OutputTokens     int64             `json:"output_tokens"`                            // 输出 Token
 	Ftut             int               `json:"ftut"`                                     // 首字时间(毫秒)
 	UseTime          int               `json:"use_time"`                                 // 总用时(毫秒)
 	Cost             float64           `json:"cost"`                                     // 消耗费用
